+++ src/constants.ts
@@ -1,11 +1,15 @@
 export const CONFIRM_MODEL_TITLE = 'Expired Notes';
-export const MESSAGE_CONFIRM_DELETION = 'The following %s note(s) have expired. Delete them now?';
-export const MESSAGE_CONFIRM_ARCHIVE = 'The following %s note(s) have expired. Archive them now?';
+
+export const MESSAGE_CONFIRM_DELETION = 'The following %s notes have expired. Delete them now?';
+export const MESSAGE_CONFIRM_DELETION_SINGULAR = 'The following note has expired. Delete it now?';
+export const MESSAGE_CONFIRM_ARCHIVE = 'The following %s notes have expired. Archive them now?';
+export const MESSAGE_CONFIRM_ARCHIVE_SINGULAR = 'The following note has expired. Archive it now?';
+
 export const BEHAVIOR_DELETE = 'delete';
 export const BEHAVIOR_ARCHIVE = 'archive';
+
 export const EXPIRY_DATE_PROMPT = 'Provide expiry date';
-export const EXPIRY_DATE_DESC = 'You can provide the date in the format %s, or you can use relative dates. Try one of the following:';
-export const EXPIRY_DATE_EXAMPLES = [
-    'today', 'tomorrow', '7 days ago', '1 month from today', 'August 12'
-];
+export const EXPIRY_DATE_DESC = 'You can provide the date in your chosen format, or you can use natural language dates. Find examples in the ';
+export const EXPIRY_DATE_LINK = 'https://github.com/joerncodes/obsidian-expiring-notes#set-expiry-date';
+export const EXPIRY_DATE_LINK_TEXT = 'GitHub documentation.';
 export const EXPIRY_DATE_PLACEHOLDER = 'Type expiry date here';